requirements.txt
colorama==0.4.6
psutil==6.1.0
Requests==2.32.3
scapy==2.6.1
SocksiPy_branch==1.01
urllib3==2.2.3
